+++ src/main/java/com/tallyto/gestorfinanceiro/core/application/services/CartaoCreditoService.java
@@ -2,7 +2,7 @@ package com.tallyto.gestorfinanceiro.core.application.services;
 
 import com.tallyto.gestorfinanceiro.core.domain.entities.CartaoCredito;
 import com.tallyto.gestorfinanceiro.core.domain.exceptions.CartaoCreditoException;
-import com.tallyto.gestorfinanceiro.core.domain.repositories.CartaoCreditoRepository;
+import com.tallyto.gestorfinanceiro.core.infra.repositories.CartaoCreditoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
