+++ src/main/java/com/tallyto/gestorfinanceiro/core/infra/repositories/ContaRepository.java
@@ -1,4 +1,4 @@
-package com.tallyto.gestorfinanceiro.core.domain.repositories;
+package com.tallyto.gestorfinanceiro.core.infra.repositories;
 
 import com.tallyto.gestorfinanceiro.core.domain.entities.Conta;
 import org.springframework.data.jpa.repository.JpaRepository;
